Split cache-keys: producer key builders -> @corpus/core/cache/keys
The corpus-agnostic half of the KV key surface moves to the shared engine so
sibling apps derive mark/enrichment keys identically. Copied VERBATIM (byte
-exact keys): keyForMark, keyForEnrichment, previousVersionKey, instanceIdOf,
recipeHash, qualifierHash, normalizeQualifier, slugTractate, slugDaf, and the
internal canonicalJSON/shortHash/slugId/pickStable/enrichmentScope.

keyForMark/keyForEnrichment took AnyMarkDefinition/AnyEnrichmentDefinition
(unions over the Talmud studio-schema/registry shapes); in core they take
minimal structural MarkKeyDef{id,cache_version} / EnrichmentKeyDef{...,scope},
which any fuller definition satisfies — so no studio dependency crosses into
core.

The Talmud SOURCE-cache helpers (keyForGemara, keyForRishonim, keyForRabbi*,
dafyomi, bridges, …) stay in packages/talmud/src/worker/cache-keys.ts, which
now imports slugDaf/slugTractate from core and RE-EXPORTS the moved builders —
so every existing './cache-keys' importer is unchanged (2-file diff).

Verified: pnpm -r typecheck (both clean), 1919 tests incl. the locked
cache-key + source-cache-keys suites (byte-exact), vite build, wrangler
deploy --dry-run. Lockfile untouched.
